added sign In component as well as firebase utils and database
