+++ binco-clothing/src/routes/navigation/navigation.component.jsx
@@ -1,0 +1,28 @@
+import { Outlet, Link } from "react-router-dom";
+import { Fragment } from "react";
+import "./navigation.styles.scss";
+
+import { ReactComponent as CrwnLogo } from "../../assets/crown.svg";
+
+const Navigation = () => {
+  return (
+    <Fragment>
+      <div className="navigation">
+        <Link className="logo-container" to="/">
+          <CrwnLogo className="logo" />
+        </Link>
+        <div className="nav-links-container">
+          <Link className="nav-link" to="/shop">
+            SHOP
+          </Link>
+          <Link className="nav-link" to="/sign-in">
+            SIGN IN
+          </Link>
+        </div>
+      </div>
+      <Outlet />
+    </Fragment>
+  );
+};
+
+export default Navigation;
